Separate MatterControlApplication to Init vs SystemWindow components
- Reduce to minimum viable SystemWindow
- Isolate initialization code in MatterControlApplication
- Prepare to collapse into discrete initialization stages
