+++ MatterControl.csproj
@@ -81,6 +81,7 @@
     <Compile Include="ActionBar\PrinterSelector.cs" />
     <Compile Include="ActionBar\ResetButton.cs" />
     <Compile Include="ActionBar\TemperatureWidgetBed.cs" />
+    <Compile Include="DesktopRootSystemWindow.cs" />
     <Compile Include="PartPreviewWindow\View3D\PrinterBar\PrintPopupMenu.cs" />
     <Compile Include="PartPreviewWindow\View3D\PrinterBar\SliceButton.cs" />
     <Compile Include="PartPreviewWindow\View3D\PrinterBar\PrinterConnectButton.cs" />
