+++ Program.cs
@@ -1,8 +1,11 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Threading;
+using MatterHackers.Agg;
 using MatterHackers.Agg.Platform;
+using MatterHackers.Agg.UI;
 using MatterHackers.MatterControl.DataStorage;
 using MatterHackers.MatterControl.SettingsManagement;
 using Mindscape.Raygun4Net;
@@ -11,17 +14,18 @@ namespace MatterHackers.MatterControl
 {
 	static class Program
 	{
-        private const int RaygunMaxNotifications = 15;
+		private const int RaygunMaxNotifications = 15;
 
-        private static int raygunNotificationCount = 0;
+		private static int raygunNotificationCount = 0;
 
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        [STAThread]
+		/// <summary>
+		/// The main entry point for the application.
+		/// </summary>
+		[STAThread]
 		public static void Main()
 		{
 			CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
+
 			Thread.CurrentThread.CurrentCulture = CultureInfo.InvariantCulture;
 
 			AggContext.Init(embeddedResourceName: "config.json");
@@ -41,7 +45,40 @@ namespace MatterHackers.MatterControl
 				AppDomain.CurrentDomain.UnhandledException += new UnhandledExceptionEventHandler(CurrentDomain_UnhandledException);
 			}
 
-			MatterControlApplication app = MatterControlApplication.Instance;
+			var timer = Stopwatch.StartNew();
+
+			// Get startup bounds from MatterControl and construct system window
+			//var systemWindow = new DesktopMainWindow(400, 200)
+			var (width, height) = DesktopRootSystemWindow.GetStartupBounds();
+
+			var systemWindow = new DesktopRootSystemWindow(width, height)
+			{
+				BackgroundColor = Color.DarkGray
+			};
+
+			AppContext.RootSystemWindow = systemWindow;
+
+			// Hook SystemWindow load and spin up MatterControl once we've hit first draw
+			systemWindow.Load += (s, e) =>
+			{
+				UiThread.RunOnIdle(() =>
+				{
+					var matterControl = new MatterControlApplication(-1, -1);
+
+					Console.WriteLine("Time to MatterControlApplication.Instance init: " + timer.Elapsed.TotalSeconds);
+
+					systemWindow.RemoveAllChildren();
+
+					systemWindow.AddChild(matterControl);
+
+					Console.WriteLine("Time to MatterControlApplication Layout: " + timer.Elapsed.TotalSeconds);
+
+					systemWindow.Invalidate();
+				});
+			};
+
+			// Block indefinitely
+			systemWindow.ShowAsSystemWindow();
 		}
 
 		private static void Application_ThreadException(object sender, ThreadExceptionEventArgs e)
@@ -64,27 +101,27 @@ namespace MatterHackers.MatterControl
 #endif
 		}
 
-        private static RaygunClient _raygunClient = GetCorrectClient();
+		private static RaygunClient _raygunClient = GetCorrectClient();
 
-        private static RaygunClient GetCorrectClient()
-        {
-            if (AggContext.OperatingSystem == OSType.Mac)
-            {
-                return new RaygunClient("qmMBpKy3OSTJj83+tkO7BQ=="); // this is the Mac key
-            }
-            else
-            {
-                return new RaygunClient("hQIlyUUZRGPyXVXbI6l1dA=="); // this is the PC key
-            }
-        }
+		private static RaygunClient GetCorrectClient()
+		{
+			if (AggContext.OperatingSystem == OSType.Mac)
+			{
+				return new RaygunClient("qmMBpKy3OSTJj83+tkO7BQ=="); // this is the Mac key
+			}
+			else
+			{
+				return new RaygunClient("hQIlyUUZRGPyXVXbI6l1dA=="); // this is the PC key
+			}
+		}
 
-        // ** Standard Winforms Main ** //
-        //[STAThread]
-        //static void Main()
-        //{
-        //	Application.EnableVisualStyles();
-        //	Application.SetCompatibleTextRenderingDefault(false);
-        //	Application.Run(new Form1());
-        //}
-    }
+		// ** Standard Winforms Main ** //
+		//[STAThread]
+		//static void Main()
+		//{
+		//	Application.EnableVisualStyles();
+		//	Application.SetCompatibleTextRenderingDefault(false);
+		//	Application.Run(new Form1());
+		//}
+	}
 }
